Add support for informix in .netcore.
