+++ dotnet/src/dotnetcore/GxClasses/GxClasses.csproj
@@ -32,6 +32,7 @@
     <Compile Include="..\..\dotnetframework\GxClasses\Data\GXDataPostgreSQL.cs" Link="Data\GXDataPostgreSQL.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Data\GXDataSqlite.cs" Link="Data\GXDataSqlite.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Data\GXDataHana.cs" Link="Data\GXDataHana.cs" />
+    <Compile Include="..\..\dotnetframework\GxClasses\Data\GXDataInformix.cs" Link="Data\GXDataInformix.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Diagnostics\Log.cs" Link="Diagnostics\Log.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Diagnostics\GXDebugManager.cs" Link="Diagnostics\GXDebugManager.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Domain\GXLDAP.cs" Link="Domain\GXLDAP.cs" />
@@ -42,6 +43,8 @@
     <Compile Include="..\..\dotnetframework\GxClasses\Domain\GXUserInfo.cs" Link="Domain\GXUserInfo.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Domain\GXWindow.cs" Link="Domain\GXWindow.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Domain\Structs.cs" Link="Domain\Structs.cs" />
+    <Compile Include="..\..\dotnetframework\GxClasses\ErrorHandling\GxADODataException.cs" Link="ErrorHandling\GxADODataException.cs" />
+    <Compile Include="..\..\dotnetframework\GxClasses\ErrorHandling\GxNotImplementedException.cs" Link="ErrorHandling\GxNotImplementedException.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Helpers\AssemblyHelper.cs" Link="Helpers\AssemblyHelper.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Domain\ExpressionEvaluator.cs" Link="Domain\ExpressionEvaluator.cs" />
     <Compile Include="..\..\dotnetframework\GxClasses\Core\GXApplication.cs" Link="Core\GXApplication.cs" />
