+++ dotnet/src/dotnetframework/GxClasses/Data/GXDataInformix.cs
@@ -13,6 +13,8 @@ using GeneXus.Configuration;
 using GeneXus.Utils;
 using System.Globalization;
 using GeneXus.Metadata;
+using GxClasses.Helpers;
+using System.IO;
 
 namespace GeneXus.Data
 {
@@ -20,7 +22,13 @@ namespace GeneXus.Data
 	{
 		static readonly ILog log = log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
 		static Assembly _ifxAssembly;
+#if NETCORE
+		internal static string InformixAssemblyName = "Informix.Net.Core";
+		const string InformixDbTypeEnum = "Informix.Net.Core.IfxType";
+#else
+		internal static string InformixAssemblyName = "IBM.Data.Informix";
 		const string InformixDbTypeEnum = "IBM.Data.Informix.IfxType";
+#endif
 		public static string SQL_NULL_DATE_10 = "0000-00-00";
 		public static string SQL_NULL_DATE_8 = "00000000";
 		private string m_serverInstance;
@@ -41,19 +49,26 @@ namespace GeneXus.Data
 				{
 					if (_ifxAssembly == null)
 					{
-						GXLogging.Debug(log, "Loading IBM.Data.Informix from GAC");
-						_ifxAssembly = Assembly.LoadWithPartialName("IBM.Data.Informix");
-						GXLogging.Debug(log, "IBM.Data.Informix Loaded from GAC");
+						string assemblyPath = Path.Combine(FileUtil.GetStartupDirectory(), $"{InformixAssemblyName}.dll");
+						GXLogging.Debug(log, $"Loading {InformixAssemblyName}.dll from:" + assemblyPath);
+#if NETCORE
+						var asl = new AssemblyLoader(FileUtil.GetStartupDirectory());
+						_ifxAssembly = asl.LoadFromAssemblyName(new AssemblyName(InformixAssemblyName));
+#else
+						GXLogging.Debug(log, $"Loading {InformixAssemblyName} from GAC");
+						_ifxAssembly = Assembly.LoadWithPartialName(InformixAssemblyName);
+						GXLogging.Debug(log, $"{InformixAssemblyName} Loaded from GAC");
+#endif
 					}
 
 				}
 				catch (Exception ex)
 				{
-					GXLogging.Error(log, "Error loading IBM.Data.Informix from GAC", ex);
+					GXLogging.Error(log, $"Error loading {InformixAssemblyName} from GAC", ex);
 				}
 				if (_ifxAssembly == null)
 				{
-					_ifxAssembly = Assembly.Load("IBM.Data.Informix");
+					_ifxAssembly = Assembly.Load(InformixAssemblyName);
 				}
 				return _ifxAssembly;
 			}
@@ -151,13 +166,13 @@ namespace GeneXus.Data
 
 		public override IDbDataParameter CreateParameter()
 		{
-			return (IDbDataParameter)ClassLoader.CreateInstance(IfxAssembly, "IBM.Data.Informix.IfxParameter");
+			return (IDbDataParameter)ClassLoader.CreateInstance(IfxAssembly, $"{InformixAssemblyName}.IfxParameter");
 		}
 		public override IDbDataParameter CreateParameter(string name, Object dbtype, int gxlength, int gxdec)
 		{
 			object ifxType = GXTypeToIfxType((GXType)dbtype);
 			object[] args = new object[] { name, ifxType, gxlength };
-			IDbDataParameter parm = (IDbDataParameter)ClassLoader.CreateInstance(IfxAssembly, "IBM.Data.Informix.IfxParameter", args);
+			IDbDataParameter parm = (IDbDataParameter)ClassLoader.CreateInstance(IfxAssembly, $"{InformixAssemblyName}.IfxParameter", args);
 			
 			ClassLoader.SetPropValue(parm, "IfxType", ifxType);
 			parm.Precision = (byte)gxdec;
@@ -186,7 +201,7 @@ namespace GeneXus.Data
 		}
 		public override DbDataAdapter CreateDataAdapeter()
 		{
-			Type odpAdapter = IfxAssembly.GetType("IBM.Data.Informix.IfxDataAdapter");
+			Type odpAdapter = IfxAssembly.GetType($"{InformixAssemblyName}.IfxDataAdapter");
 			return (DbDataAdapter)Activator.CreateInstance(odpAdapter);
 		}
 
@@ -542,7 +557,7 @@ namespace GeneXus.Data
 			try
 			{
 				GXLogging.Debug(log, "Creating Informix data provider ");
-				_connection = (IDbConnection)ClassLoader.CreateInstance(GxInformix.IfxAssembly, "IBM.Data.Informix.IfxConnection");
+				_connection = (IDbConnection)ClassLoader.CreateInstance(GxInformix.IfxAssembly, $"{GxInformix.InformixAssemblyName}.IfxConnection");
 				GXLogging.Debug(log, "Informix data provider created");
 			}
 			catch (Exception ex)
@@ -556,7 +571,7 @@ namespace GeneXus.Data
 		{
 			try
 			{
-				_connection = (IDbConnection)ClassLoader.CreateInstance(GxInformix.IfxAssembly, "IBM.Data.Informix.IfxConnection", new object[] { connectionString });
+				_connection = (IDbConnection)ClassLoader.CreateInstance(GxInformix.IfxAssembly, $"{GxInformix.InformixAssemblyName}.IfxConnection", new object[] { connectionString });
 				m_isolationLevel = isolationLevel;
 				m_connectionCache = connCache;
 			}
