+++ dotnet/src/dotnetframework/GxClasses/ErrorHandling/GxNotImplementedException.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Runtime.Serialization;
 using System.Text;
@@ -8,7 +8,7 @@ namespace GeneXus.Data
 	[Serializable()]
 	public class GxNotImplementedException : Exception
 	{
-		public GxNotImplementedException(SerializationInfo info, StreamingContext ctx) : base()
+		public GxNotImplementedException(SerializationInfo info, StreamingContext ctx) : base(info, ctx)
 		{
 
 		}
